chore: deprecated linkNativeMethod
