+++ packages/web/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./error";
+export * from "./linkBridge";
 export * from "./linkNativeMethod";
 export * from "./registerWebMethod";
 export type * from "./types";
